fix: correct indicator center coordinates and move helper out of symbols/
The (cx, cy) values passed to indicator path functions were wrong because
they assumed rotation around the origin, but rotateRightFacingSymbol
rotates around the port midpoint (0, 0.035). Computed correct post-rotation
circle centers:
  - up/vert: (-0.005, 0.035)
  - down: (0.005, 0.035)
  - left: (0, 0.03)

Also moved voltmeter-indicator-paths.ts from symbols/ to drawing/ so the
VFS generator (make-vfs) doesn't try to import it as a symbol with a
default export, which broke the build.
