+++ drawing/voltmeter-indicator-paths.ts
