+++ symbols/ac_voltmeter_left.ts
@@ -1,7 +1,7 @@
 import { modifySymbol } from "../drawing/modify-symbol/modify-symbol"
 import type { Primitive } from "../drawing/types"
 import svgJson from "assets/generated/ac_voltmeter.json"
-import { getACVoltmeterIndicatorPaths } from "./voltmeter-indicator-paths"
+import { getACVoltmeterIndicatorPaths } from "../drawing/voltmeter-indicator-paths"
 
 const { paths, texts, bounds, refblocks, circles } = svgJson
 
@@ -17,10 +17,10 @@ const {
   ...structuralPaths
 } = paths
 
-// After 180° rotation the circle center moves from (0, 0.04) to (0, -0.04).
-// Decorative paths keep their original shape, translated to the new center.
+// After 180° rotation (around port midpoint (0, 0.035)) the circle center
+// moves from (0, 0.04) to (0, 0.03).
 const cx = 0
-const cy = -0.04
+const cy = 0.03
 
 export default modifySymbol({
   primitives: [
